requirements.txt
google-generativeai
streamlit
firecrawl-py